requirements.txt
aiogram==3.10.0
aiohttp==3.9.1
python-dotenv==1.0.0
pydantic==2.8.0
aiofiles==23.2.1
aiosqlite==0.19.0
structlog==23.2.0
openai==1.40.0
anthropic==0.34.0
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0
black==23.12.0
flake8==7.0.0
mypy==1.8.0
cryptography==46.0.1
